paper/main.tex: Incorporate general peer-review
Tweak the introduction to hopefully make it more clear what is being
addressed in this research. Criticize other datasets more and outline
why the RMSE was not chosen. Further, better describe the process of
calculating the mean recall-position in a Top-N set.

Furthermore, drop supervisors except for Andrew, i.e. drop Joeran and
Akiko from the list of authors. Change the repository link to point to
GitHub:Edenhofer instead of GitHub:BeelGroup.

Additionally fix some typos and marginally shrink the space between
figures and text.